+++ Editor/SpriteSelector.cs
@@ -1,10 +1,5 @@
 ﻿using System;
-using System.IO;
-using System.Collections.Generic;
-using System.Linq;
 using UnityEngine;
-using UnityEditorInternal;
-using UnityEditor.Experimental;
 
 namespace UnityEditor.U2D
 {